按照规范将ScriptModule定义合并至__init__中，并修改部分小问题
